Bei dao poems t00065 t00069 (#70)
* Bei-Dao-poems-T00065-T00069

added poems to the text file and publication place for some entries. Therefore I created a new entry in the place file.

I have difficulties to find out the english titles for some poems. See #67
Maybe Chia-yi could help me with this?

* title-T00069

I added the english title for T00069 and made 2 entries in the Act file on T00065 and T00066